movig scraper logic to utils/scraper
